add loading, login, delete productInCart, fix error discount percentage

+++ client/src/context/useCart.jsx
@@ -1,27 +1,47 @@
 import { createContext, useState } from "react";
 import ProductService from '../service/ProductService';
 import CartService from '../service/CartService';
+import { Alert } from "react-native";
 
 export const CartContext = createContext();
 
 export function CartProvider({children}) {
-    const [token, setToken] = useState("eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiIyIiwiaWF0IjoxNjk2Mjg1NzE4LCJzdWIiOiJjYXJsb3MubWFydGluZXpAZW1haWwuY29tIiwiaXNzIjoiTWFpbiIsImV4cCI6MTY5Njg5MDUxOH0.Ni12tYb2et0gjznvZRgAtIygVV9jHd4NLhiF2IlzLYU");
+    const [isSignedIn, setIsSignedIn] = useState(false);
+    const [token, setToken] = useState("");
     const [products, setProducts] = useState([]);
     const [cart, setCart] = useState([]);
+    const [loadingProducts, setLoadingProducts] = useState(true);
+    const [loadingCart, setLoadingCart] = useState(true);
+    const [totalPrice, setTotalPrice] = useState(0);
+
+    const totalPriceCalculator = () => {
+        const total = cart.reduce((acc, product)=> {
+            const priceWithDiscount = product.price * (1 - product.discountPercentage/100);
+            return acc + (priceWithDiscount * product.quantity);
+        }, 0); 
+        setTotalPrice(total);
+    }
 
     const listProducts = () => {
         ProductService.getAllProducts(token).then(response=> {
             setProducts(response.data);
         }).catch(error=> {
             console.log(error);
+        }).finally(()=> {
+            setLoadingProducts(false);
         })
     }
 
     const addToCart = (productId) => {
         CartService.addProductInCart(token, productId).then(response=> {
             cartList();
+            totalPriceCalculator();
         }).catch(error=> {
-            console.log(error);
+            if(error.response && error.response.status === 400) {
+                alert(error.response.data);
+            } else {
+                console.log(error);
+            }
         })
     }
 
@@ -30,6 +50,7 @@ export function CartProvider({children}) {
         if(cart[productInCartIndex].quantity  > 1) {
             const newCart = [...cart]
             newCart[productInCartIndex].quantity -= 1
+            totalPriceCalculator();
             return setCart(newCart)
         }
     }
@@ -39,12 +60,21 @@ export function CartProvider({children}) {
         if(cart[productInCartIndex].quantity  < 15) {
             const newCart = [...cart]
             newCart[productInCartIndex].quantity += 1
+            totalPriceCalculator();
             return setCart(newCart)
         }
     }
 
     const clearCart = () => {
-        setCart([])
+        Alert.alert('Confirm Checkout', 'Are you sure you want to proceed with the checkout?', [
+            {text: 'Cancel'},
+            {
+                text: 'Yes',
+                onPress: ()=> {
+                    console.log("checkout");
+                }
+            }
+        ])
     }
 
     const cartList = () => {
@@ -57,15 +87,26 @@ export function CartProvider({children}) {
             setCart(productData);
         }).catch(error=> {
             console.log(error);
+        }).finally(()=> {
+            setLoadingCart(false);
         })
     }
 
     const removeFromCart = (cartId) => {
-        CartService.deleteProductInCart(token, cartId).then(response=> {
-            cartList();
-        }).catch(error=> {
-            console.log(error);
-        })
+        Alert.alert('Delete Product', 'Are you sure you want to delete product from cart?', [
+            {text: 'Cancel'},
+            {
+                text: 'Yes',
+                onPress: ()=> {
+                    CartService.deleteProductInCart(token, cartId).then(response=> {
+                        cartList();
+                        totalPriceCalculator();
+                    }).catch(error=> {
+                        console.log(error);
+                    })
+                }
+            }
+        ])
     }
 
 
@@ -79,7 +120,14 @@ export function CartProvider({children}) {
             cartList,
             listProducts,
             products,
-            cart
+            cart,
+            loadingProducts,
+            loadingCart,
+            isSignedIn,
+            setToken,
+            setIsSignedIn,
+            totalPrice,
+            totalPriceCalculator
         }} >
             {children}
         </CartContext.Provider>
